moving fixtures to conftest
