+++ src/konnektor/tests/network_planners/concatenators/test_bipartite_MST_connector.py
@@ -7,9 +7,7 @@ from konnektor.network_analysis import get_is_connected
 from konnektor.network_planners.concatenators import MstConcatenator
 from konnektor.tests.network_planners.conf import (
     GenAtomMapper,
-    atom_mapping_basic_test_files,
     genScorer,
-    ligand_network_ab,
 )
