+++ src/konnektor/tests/network_planners/concatenators/test_bipartite_max_concatenator.py
@@ -7,9 +7,7 @@ from gufe import LigandNetwork
 from konnektor.network_planners.concatenators.max_concatenator import MaxConcatenator
 from konnektor.tests.network_planners.conf import (
     GenAtomMapper,
-    atom_mapping_basic_test_files,
     genScorer,
-    ligand_network_ab,
 )
