+++ src/konnektor/tests/network_planners/generators/test_maximal_network_planner.py
@@ -2,18 +2,14 @@
 # For details, see https://github.com/OpenFreeEnergy/konnektor
 
 import pytest
-from gufe import AtomMapper, AtomMapping, LigandAtomMapping
 
 from konnektor.network_planners import MaximalNetworkGenerator
 from konnektor.tests.network_planners.conf import (
     BadMapper,
     GenAtomMapper,
     SuperBadMapper,
-    atom_mapping_basic_test_files,
     genScorer,
-    toluene_vs_others,
 )
-from konnektor.utils.toy_data import build_random_dataset
 
 
 @pytest.mark.parametrize("n_process", [1, 2])
