+++ src/konnektor/tests/network_planners/generators/test_star_network_planner.py
@@ -9,10 +9,8 @@ from konnektor.tests.network_planners.conf import (
     BadMapper,
     ErrorMapper,
     GenAtomMapper,
-    atom_mapping_basic_test_files,
     genScorer,
     mol_from_smiles,
-    toluene_vs_others,
 )
